requirements.txt
anyio==4.3.0
APScheduler==3.10.4
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
mypy_extensions==0.4.4
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
toml==0.10.2
types-requests==2.31.0.20240218
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
